refactor: update to hexagonal architecture
