+++ books-service/internal/adapters/database/dbConfig.go
@@ -1,4 +1,4 @@
-package config
+package database
 
 import (
 	"fmt"
