+++ books-service/internal/adapters/database/pinger.go
