+++ books-service/internal/adapters/database/sqlite.go
