+++ books-service/internal/adapters/http/dtos/responseDTO.go
