+++ books-service/internal/adapters/http/server.go
@@ -1,4 +1,4 @@
-package application
+package http
 
 import (
 	"log"
